Refactor lib interface from Idr to CveApi
The CVE Services API consists of several "modules" of functionality:

- IDR: CVE ID reservation service
- User Registry:

The initial version of this lib and CLI commands only supported the
existing IDR functionality and assumed the other two "modules" would be
somehow split across several APIs. The current architecture of the CVE
Services API is a single REST API with specific endpoints that support the
different pieces of functionality. Thus, it makes sense to expose the
entire API as a single interface with supporting methods for each piece
of functionality.

This commit refactors the CLI commands, env variables, and lib interface to
remove any mention of IDR and present the entire CVE Services API as a
single interface.