Remove the "owning CNA" option when reserving IDs
This action is limited to the MITRE Secretariat role. Since MITRE is not a
user of this client, allowing reservation of IDs for other CNAs is an
unused piece of functionality.

This functionality can be added back in the future when more organizations
than just MITRE will be allowed to have administrative privileges.