Add a few missing state transitions
This adds support for moving CVE IDs to rejected without a CVE record as
well as moving such rejected IDs back to reserved. For lack of a better
action name, the subcommand is called undo-reject since the only transition
it supports is moving a rejected w/o record ID to reserved (which too has
no record).

This also adds a convenience Constants enum to use when we need lists of
constant values.

Misc improvements to docs and options as well.