Add ADP functionality
A new `publish-adp` command is added that allows publishing of ADP
containers into an existing CVE record (this is only possible if a CVE is
in the published state).

The `show` subcommand now allows displaying a CNA container or all/subset
of existing ADP containers (identified by the org's name that created it).

ADP containers can only be published and updated, so there is no
functionality to remove them.